Initial support for deployment of remotely stored images. Remote images are exported from remote into local bravetools imagestore before deployment. Issues occur when two-part image files are used.
